+++ packages/flame/lib/src/components/mixins/hoverable.dart
@@ -16,18 +16,19 @@ mixin Hoverable on Component {
   }
 
   @nonVirtual
-  void handleMouseMovement(PointerHoverInfo info) {
+  bool handleMouseMovement(PointerHoverInfo info) {
     if (containsPoint(eventPosition(info))) {
       if (!_isHovered) {
         _isHovered = true;
-        onHoverEnter(info);
+        return onHoverEnter(info);
       }
     } else {
       if (_isHovered) {
         _isHovered = false;
-        onHoverLeave(info);
+        return onHoverLeave(info);
       }
     }
+    return true;
   }
 
   @override
@@ -49,8 +50,7 @@ mixin HasHoverableComponents on FlameGame {
   @mustCallSuper
   void onMouseMove(PointerHoverInfo info) {
     bool _mouseMoveHandler(Hoverable c) {
-      c.handleMouseMovement(info);
-      return true; // always continue
+      return c.handleMouseMovement(info);
     }
 
     for (final c in children.reversed()) {