+++ packages/flame/CHANGELOG.md
@@ -1,7 +1,6 @@
 # CHANGELOG
 
 ## [Next]
- - Changed `onHoverEnter` and `onHoverLeave` to return `bool` (breaking change)
  - `changePriority` no longer breaks game loop iteration
  - Move component mixin checks to their own files
  - Fix exception when game was rebuilt
@@ -9,6 +8,7 @@
  - Add `SpriteSheet.createAnimationVariable` method to allow animations with different `stepTime` for each sprite
  - Use the full delta in `JoystickComponent` so that it can't go to the wrong direction on the wrong side
  - Introduce `onDoubleTapDown` with info and `onDoubleTapCancel`
+ - Changed `onHoverEnter` and `onHoverLeave` to return `bool` (breaking change)
 
 ## [1.0.0-releasecandidate.15]
  - Fix issue with `Draggable`s not being removed from `draggables` list